affichage des produit, d'un produit et ajoute au panier

+++ ecommerce/src/main/java/fr/orleans/m1/wsi/ecommerce/models/Product.java
@@ -28,7 +28,7 @@ public class Product {
     private String description;
 
     @Column(nullable = false)
-    private long stock = 0;
+    private Integer stock = 0;
 
     private boolean active = true;
 }
